+++ src/c2d2-browser.js
@@ -1,0 +1,5 @@
+import c2d2 from './c2d2.js';
+
+c2d2.setWindow(window);
+
+export default c2d2;